+++ packages/eslint-config/index.js
@@ -2,8 +2,7 @@ module.exports = {
     extends: [
         'eslint-config-airbnb',
         '@paysera/eslint-config-base',
+        './rules/react',
     ].map(require.resolve),
-    rules: {
-        strict: ['error', 'never']
-    }
+    rules: {}
 };