+++ packages/eslint-config-base/CHANGELOG.md
@@ -1,4 +1,4 @@
 1.0.0 / 2018-08-16
 ==================
-    - [breaking] update `eslint-config-airbnb-base` to v13 (https://github.com/airbnb/javascript/blob/master/packages/eslint-config-airbnb-base/CHANGELOG.md)
-    - [patch] `class-methods-use-this`: set to "off"
+  - [breaking] update `eslint-config-airbnb-base` to v13 (https://github.com/airbnb/javascript/blob/master/packages/eslint-config-airbnb-base/CHANGELOG.md)
+  - [patch] `class-methods-use-this`: set to "off"